+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.143.0 - 26.04.2022
+
+### Changed
+
+- Improved the filtering by tags
+
 ## 1.142.0 - 25.04.2022
 
 ### Added